Add BCDD::Resultable
This module can add Success() and Failure() in any object. The target object will be
the subject of the result object produced by these methods.

Make the BCDD::Result enabled to receive a subject.

Make the BCDD::Result#and_then method receive a method name (symbol),
and perform it on the result subject.